requirements.txt
openai
aiohttp
python-dotenv
discord.py
apscheduler
